--- a/vllm/v1/core/scheduler.py
+++ b/vllm/v1/core/scheduler.py
@@ -210,7 +210,6 @@ class Scheduler:
         )
 
         self.finished_req_ids = set()
-        self.kv_cache_manager.async_remove_touched_blocks()
         return scheduler_output
 
     def _make_running_request_data(