@comaniac 2024-11-07
> A high level question: Right now for a block, we have 3 ways to index it: block hash, block id, and the python `KVCacheBlock` object itself. Do we have to have all 3? I assume we cannot just use block hash since not all blocks have hash. Can we remove block id and only use `KVCacheBlock` object for index? Is the reason we keep `block_id` is that we need to pass `block_id`s to the worker?

So the question is about whether we could remove "block_id" attribute from the data class? I'm not sure but will take a look.