@comaniac 2024-11-08
The changes in the latest commit:
1. Refactor utilities to `kv_cache_utils.py`.
2. Enhance the block hash type from `int` to `Tuple[int, Tuple[int]]`, which is (hash_value, (toke_ids,)). This guarantees no hash conflicts. I benchmarked the tuple matching latency and it is about 0.02 ms for block size 16,32,48. It is ~3x faster than list matching.
3. Use `request.all_token_ids` and remove "prefill"/"decode" specific logic.
4. Enable prefix caching by default in v1.